Implement route-level code splitting and lazy loading
