+++ src/app/cex/page.tsx
@@ -1,147 +1,9 @@
-"use client";
+import dynamic from "next/dynamic";
 
-import { useState } from "react";
-import { Building2, Copy, Check, ExternalLink, Wallet } from "lucide-react";
-import { CEX_LIST } from "@/lib/types";
+const CexPage = dynamic(() => import("@/components/routes/cex-page"), {
+  loading: () => <div className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8 py-24 text-sm text-[var(--text-muted)]">Loading CEX route…</div>,
+});
 
-export default function CexPage() {
-  const [selectedCex, setSelectedCex] = useState(CEX_LIST[0]);
-  const [cAddress, setCAddress] = useState("");
-  const [copiedField, setCopiedField] = useState<string | null>(null);
-
-  const handleCopy = (text: string, field: string) => {
-    navigator.clipboard.writeText(text);
-    setCopiedField(field);
-    setTimeout(() => setCopiedField(null), 2000);
-  };
-
-  const withdrawalUrl = selectedCex.withdrawalUrl;
-
-  return (
-    <div className="max-w-4xl mx-auto px-4 sm:px-6 lg:px-8 py-12">
-      <div className="mb-8">
-        <h1 className="text-3xl font-bold mb-2">CEX Withdrawal Routing</h1>
-        <p className="text-[var(--text-muted)]">
-          Route your centralized exchange withdrawals directly to a Soroban C-address.
-        </p>
-      </div>
-
-      <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
-        <div className="lg:col-span-2 space-y-6">
-          <div className="rounded-xl border border-[var(--border)] bg-[var(--surface)] p-6">
-            <h3 className="font-semibold mb-4">1. Select Your Exchange</h3>
-            <div className="grid grid-cols-1 sm:grid-cols-3 gap-3">
-              {CEX_LIST.map((cex) => (
-                <button
-                  key={cex.name}
-                  onClick={() => setSelectedCex(cex)}
-                  className={`p-4 rounded-lg border text-left transition-all ${
-                    selectedCex.name === cex.name
-                      ? "border-[var(--primary)] bg-[var(--primary)]/5"
-                      : "border-[var(--border)] bg-[var(--surface-2)] hover:border-[var(--text-muted)]"
-                  }`}
-                >
-                  <Building2 className="w-8 h-8 text-[var(--text-muted)] mb-2" />
-                  <div className="font-medium text-sm">{cex.name}</div>
-                  <div className="text-xs text-[var(--text-muted)]">Min: {cex.minWithdrawal}</div>
-                </button>
-              ))}
-            </div>
-          </div>
-
-          <div className="rounded-xl border border-[var(--border)] bg-[var(--surface)] p-6">
-            <h3 className="font-semibold mb-4">2. Enter Your C-Address</h3>
-            <p className="text-xs text-[var(--text-muted)] mb-3">
-              This is the Soroban smart account address you want to fund.
-              It will be linked to your deposit via the bridge memo.
-            </p>
-            <div className="relative">
-              <Wallet className="absolute left-3 top-1/2 -translate-y-1/2 w-4 h-4 text-[var(--text-muted)]" />
-              <input
-                type="text"
-                value={cAddress}
-                onChange={(e) => setCAddress(e.target.value)}
-                placeholder="CABC...DEF"
-                className="w-full pl-10 pr-4 py-3 rounded-lg bg-[var(--surface-2)] border border-[var(--border)] text-sm font-mono focus:outline-none focus:border-[var(--primary)] transition-colors"
-              />
-            </div>
-          </div>
-
-          <div className="rounded-xl border border-[var(--border)] bg-[var(--surface)] p-6">
-            <h3 className="font-semibold mb-4">3. Withdrawal Details for {selectedCex.name}</h3>
-            <div className="space-y-4">
-              {cAddress && (
-                <div>
-                  <label className="block text-xs text-[var(--text-muted)] mb-1">Your C-Address (for reference)</label>
-                  <div className="flex items-center gap-2">
-                    <code className="flex-1 p-3 rounded-lg bg-[var(--surface-2)] border border-[var(--border)] text-xs font-mono break-all">
-                      {cAddress}
-                    </code>
-                    <button
-                      onClick={() => handleCopy(cAddress, "caddress")}
-                      className="p-3 rounded-lg border border-[var(--border)] hover:bg-[var(--surface-2)] transition-colors"
-                    >
-                      {copiedField === "caddress" ? <Check className="w-4 h-4 text-[var(--success)]" /> : <Copy className="w-4 h-4 text-[var(--text-muted)]" />}
-                    </button>
-                  </div>
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-
-        <div className="space-y-4">
-          <div className="rounded-xl border border-[var(--border)] bg-[var(--surface)] p-5">
-            <h3 className="font-semibold mb-3">Exchange Details</h3>
-            <div className="space-y-2 text-sm">
-              <div className="flex justify-between">
-                <span className="text-[var(--text-muted)]">Exchange</span>
-                <span>{selectedCex.name}</span>
-              </div>
-              <div className="flex justify-between">
-                <span className="text-[var(--text-muted)]">Min Withdrawal</span>
-                <span>{selectedCex.minWithdrawal}</span>
-              </div>
-              <div className="flex justify-between">
-                <span className="text-[var(--text-muted)]">Fee</span>
-                <span>{selectedCex.fee}</span>
-              </div>
-            </div>
-          </div>
-
-          <div className="rounded-xl border border-[var(--border)] bg-[var(--surface)] p-5">
-            <h3 className="font-semibold mb-3">How It Works</h3>
-            <ol className="space-y-3 text-sm text-[var(--text-muted)]">
-              <li className="flex gap-2">
-                <span className="text-[var(--primary-light)] font-medium">1.</span>
-                <span>Withdraw from your CEX to the bridge address</span>
-              </li>
-              <li className="flex gap-2">
-                <span className="text-[var(--primary-light)] font-medium">2.</span>
-                <span>The Soroban bridge contract detects the deposit</span>
-              </li>
-              <li className="flex gap-2">
-                <span className="text-[var(--primary-light)] font-medium">3.</span>
-                <span>Funds are routed to your C-address automatically</span>
-              </li>
-              <li className="flex gap-2">
-                <span className="text-[var(--primary-light)] font-medium">4.</span>
-                <span>Use your Soroban dApp directly</span>
-              </li>
-            </ol>
-          </div>
-
-          <a
-            href={withdrawalUrl}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="flex items-center justify-center gap-2 px-4 py-3 rounded-xl border border-[var(--border)] text-sm font-medium hover:bg-[var(--surface-2)] transition-colors"
-          >
-            <ExternalLink className="w-4 h-4" />
-            Open {selectedCex.name} Withdrawal
-          </a>
-        </div>
-      </div>
-    </div>
-  );
+export default function CexRoutePage() {
+  return <CexPage />;
 }
