+++ src/lib/stellar.ts
@@ -4,15 +4,7 @@ import {
   signTransaction,
   getNetwork,
 } from "@stellar/freighter-api";
-import {
-  TransactionBuilder,
-  Operation,
-  BASE_FEE,
-  Networks,
-  Asset,
-  Horizon,
-  rpc,
-} from "@stellar/stellar-sdk";
+import type { Horizon, rpc } from "@stellar/stellar-sdk";
 import { BRIDGE_CONTRACT_ID } from "./types";
 
 const HORIZON_URLS = {
@@ -25,15 +17,18 @@ const SOROBAN_RPC_URLS = {
   TESTNET: "https://soroban-rpc-testnet.stellar.org",
 };
 
-export function getHorizonServer(network: "PUBLIC" | "TESTNET"): Horizon.Server {
+export async function getHorizonServer(network: "PUBLIC" | "TESTNET"): Promise<Horizon.Server> {
+  const { Horizon } = await import("@stellar/stellar-sdk");
   return new Horizon.Server(HORIZON_URLS[network]);
 }
 
-export function getSorobanRpcServer(network: "PUBLIC" | "TESTNET"): rpc.Server {
+export async function getSorobanRpcServer(network: "PUBLIC" | "TESTNET"): Promise<rpc.Server> {
+  const { rpc } = await import("@stellar/stellar-sdk");
   return new rpc.Server(SOROBAN_RPC_URLS[network]);
 }
 
-export function getNetworkPassphrase(network: "PUBLIC" | "TESTNET"): string {
+export async function getNetworkPassphrase(network: "PUBLIC" | "TESTNET"): Promise<string> {
+  const { Networks } = await import("@stellar/stellar-sdk");
   return network === "PUBLIC" ? Networks.PUBLIC : Networks.TESTNET;
 }
 
@@ -72,7 +67,8 @@ export async function getWalletAddress(): Promise<string | null> {
 export async function getCurrentNetwork(): Promise<"PUBLIC" | "TESTNET"> {
   try {
     const result = await getNetwork();
-    return result.network === Networks.PUBLIC ? "PUBLIC" : "TESTNET";
+    const networkName = String(result.network ?? "").toUpperCase();
+    return networkName === "PUBLIC" ? "PUBLIC" : "TESTNET";
   } catch {
     return "TESTNET";
   }
@@ -136,7 +132,7 @@ export async function getAccountBalances(
   address: string,
   network: "PUBLIC" | "TESTNET"
 ): Promise<AccountBalances> {
-  const server = getHorizonServer(network);
+  const server = await getHorizonServer(network);
   try {
     const account = await server.loadAccount(address);
     const balances = (account.balances as HorizonBalance[]).map((b) => ({
@@ -155,7 +151,7 @@ export async function fetchRecentTransactions(
   network: "PUBLIC" | "TESTNET",
   limit: number = 10
 ): Promise<BridgeTransactionData[]> {
-  const server = getHorizonServer(network);
+  const server = await getHorizonServer(network);
   try {
     const payments = await server
       .payments()
@@ -187,11 +183,13 @@ export async function buildAndSubmitPayment(
   assetCode: string,
   network: "PUBLIC" | "TESTNET"
 ): Promise<PaymentResult> {
-  const server = getHorizonServer(network);
-  const passphrase = getNetworkPassphrase(network);
+  const server = await getHorizonServer(network);
+  const passphrase = await getNetworkPassphrase(network);
+  const { TransactionBuilder, Operation, BASE_FEE, Asset } = await import("@stellar/stellar-sdk");
+  type AssetType = InstanceType<typeof Asset>;
 
   const account = await server.loadAccount(sourceAddress);
-  let asset: Asset;
+  let asset: AssetType;
   if (assetCode === "XLM") {
     asset = Asset.native();
   } else {
@@ -249,8 +247,9 @@ export async function bridgeViaContract(
     return buildAndSubmitPayment(sourceAddress, cAddress, amount, assetCode, network);
   }
 
-  const server = getHorizonServer(network);
-  const passphrase = getNetworkPassphrase(network);
+  const server = await getHorizonServer(network);
+  const passphrase = await getNetworkPassphrase(network);
+  const { TransactionBuilder, Operation, BASE_FEE, Asset } = await import("@stellar/stellar-sdk");
 
   const account = await server.loadAccount(sourceAddress);
